feat: Modify controller and logic for Parking

+++ src/main/java/com/tutorial/crud/repository/ParqueaderoRepository.java
@@ -3,5 +3,5 @@ package com.tutorial.crud.repository;
 import com.tutorial.crud.entity.Parqueadero;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParqueaderoRepository extends JpaRepository<Parqueadero, Long> {
+public interface ParqueaderoRepository extends JpaRepository<Parqueadero, Integer> {
 }
